CAMBIAMOS SESSION

+++ controllers/productController.js
@@ -6,15 +6,14 @@ const productsFilePath = path.join(__dirname, "../data/products.json");
 
 const { validationResult } = require("express-validator"); //LLAMO AL OBJETO DE ERRORES DE VALIDACIONES
 
-
 let controller = {
     
     productDetail: function (req,res){
 
         /*RENDERIZO LA VISTA DEL DETALLE DEL PRODUCTO DE ACUERDO AL ID SELECCIONADO*/
         let productos = JSON.parse(fs.readFileSync(productsFilePath, "utf-8"));
-        producto = productos.find((p) => p.id == req.params.id);
-        res.render("product-detail", { producto: producto,loguin:req.session.usuarioLogueado });
+        let producto = productos.find((p) => p.id == req.params.id);
+        res.render("product-detail", { producto: producto});
 
         },  
     carro: function(req,res){
@@ -22,34 +21,21 @@ let controller = {
          },
 
     crear: function(req,res){
-
-        if(req.session.usuarioLogueado){
-
-            res.render("crear", {loguin:req.session.usuarioLogueado});
-            /*RENDERIZO LA VISTA DEL FORMULARIO DE CREAR*/
-
-        }else{
-            res.redirect("/user/login")
-        }
+        /*RENDERIZO LA VISTA DEL FORMULARIO DE CREAR*/
+    res.render("crear");
          },
 
     editar: function(req,res){
-
-        if(req.session.usuarioLogueado){
         /*LEO EL ARCHIVO Y RENDERIZO LA VISTA DE EDITAR DE PRODUCTOS*/
         const productos = JSON.parse(fs.readFileSync(productsFilePath, "utf-8"));
         const producto = productos.find((p) => p.id == req.params.id);
-        res.render("editar", { producto: producto, loguin:req.session.usuarioLogueado});
-        
-         } else{
-            res.redirect("/user/login")
-         }
+        res.render("editar", {producto: producto});
         },
     lista: function(req,res){
 
         /*LEO EL ARCHIVO Y RENDERIZO LA VISTA DE LISTA DE PRODUCTOS*/
         const productos = JSON.parse(fs.readFileSync(productsFilePath, "utf-8"));
-        res.render("lista", {productos:productos ,loguin:req.session.usuarioLogueado}); 
+        res.render("lista", {productos:productos}); 
         },
     guardar: function(req,res){
         const productos = JSON.parse(fs.readFileSync(productsFilePath, "utf-8"));
@@ -95,7 +81,7 @@ let controller = {
             p.marca =req.body.marca;
 
         /*VEO SI HAY ARCHIVO Y CAMBIO NOMBRE DE IMAGEN GUARDADA*/
-        if (req.file) {
+        if (req.file && req.body.image !="user-vacio.jpg") {
             fs.unlinkSync("./public/bancoimagenes/" + p.image);
             p.image = req.file.filename;
             }
@@ -110,10 +96,9 @@ let controller = {
     res.redirect("/products/product-detail/" + req.params.id);
     
      },
-     
+
     borrar: function(req,res){
     
-    if(req.session.usuarioLogueado){
     /*LEO EL ARCHIVO DE PRODUCTOS*/
     let productos = JSON.parse(fs.readFileSync(productsFilePath, "utf-8"));
     /*BUSCO ID QUE VINO POR PARAMETRO*/
@@ -128,9 +113,6 @@ let controller = {
     const data = JSON.stringify(productos, null, " ");
     fs.writeFileSync(productsFilePath, data);
     res.redirect("/");
-    }else{
-        res.redirect("/user/login")
-        }
     }
 }
